Cleaned up trajectory generators and switched them all to using switch case statements instead of if else statements.

+++ iam_robolib/src/trajectory_generator/linear_joint_trajectory_generator.cpp
@@ -7,7 +7,7 @@
 void LinearJointTrajectoryGenerator::get_next_step() {
   t_ = std::min(std::max(time_ / run_time_, 0.0), 1.0);
 
-  for (int i = 0; i < desired_joints_.size(); i++) {
+  for (size_t i = 0; i < desired_joints_.size(); i++) {
     desired_joints_[i] = initial_joints_[i] * (1 - t_) + goal_joints_[i] * t_;
   }
 }
